armbian-next: manual merge (14) of all changes between revisions 13469fd8a9eda4a53f87b7dfee0f431dbab8db54 and 09e416e31cc01ece4533a65f02a470a4c21b90ea
- also editorconfig and compile.sh (root) changes

+++ lib/functions/cli/utils-cli.sh
@@ -111,7 +111,7 @@ function prepare_userpatches() {
 	mkdir -p "${SRC}"/userpatches
 
 	# Create example configs if none found in userpatches
-	if ! ls "${SRC}"/userpatches/{config-example.conf,config-docker.conf,config-vagrant.conf} 1> /dev/null 2>&1; then
+	if ! ls "${SRC}"/userpatches/{config-default.conf,config-docker.conf,config-vagrant.conf} 1> /dev/null 2>&1; then
 
 		# Migrate old configs
 		if ls "${SRC}"/*.conf 1> /dev/null 2>&1; then
@@ -126,6 +126,10 @@ function prepare_userpatches() {
 		# Create example config
 		if [[ ! -f "${SRC}"/userpatches/config-example.conf ]]; then
 			cp "${SRC}"/config/templates/config-example.conf "${SRC}"/userpatches/config-example.conf || exit 1
+		fi
+
+		# Link default config to example config
+		if [[ ! -f "${SRC}"/userpatches/config-default.conf ]]; then
 			ln -fs config-example.conf "${SRC}"/userpatches/config-default.conf || exit 1
 		fi
